requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
fake-useragent==0.0.8
idna==2.10
lxml==4.5.2
mysql-connector-python==8.0.21
pkg-resources==0.0.0
protobuf==3.12.2
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
